Change HIDE_GENERATION_TIMESTAMP handling
