+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/dart/DartClientCodegenTest.java
@@ -13,7 +13,7 @@ public class DartClientCodegenTest {
         final DartClientCodegen codegen = new DartClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
