+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/eiffel/EiffelClientCodegenTest.java
@@ -13,7 +13,7 @@ public class EiffelClientCodegenTest {
         final EiffelClientCodegen codegen = new EiffelClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
