+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/go/GoClientCodegenTest.java
@@ -13,7 +13,7 @@ public class GoClientCodegenTest {
         final GoClientCodegen codegen = new GoClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
