+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/javascript/JavascriptClosureAngularClientCodegenTest.java
@@ -13,7 +13,7 @@ public class JavascriptClosureAngularClientCodegenTest {
         final JavascriptClosureAngularClientCodegen codegen = new JavascriptClosureAngularClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
