+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/lua/LuaClientCodegenTest.java
@@ -13,7 +13,7 @@ public class LuaClientCodegenTest {
         final LuaClientCodegen codegen = new LuaClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
