+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/objc/ObjcClientCodegenTest.java
@@ -13,7 +13,7 @@ public class ObjcClientCodegenTest {
         final ObjcClientCodegen codegen = new ObjcClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
