+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/perl/PerlClientCodegenTest.java
@@ -13,7 +13,7 @@ public class PerlClientCodegenTest {
         final PerlClientCodegen codegen = new PerlClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
