+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/php/PhpClientCodegenTest.java
@@ -13,7 +13,7 @@ public class PhpClientCodegenTest {
         final PhpClientCodegen codegen = new PhpClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
