+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/phpsymfony/SymfonyServerCodegenTest.java
@@ -13,7 +13,7 @@ public class SymfonyServerCodegenTest {
         final SymfonyServerCodegen codegen = new SymfonyServerCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
