+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/r/RClientCodegenTest.java
@@ -13,7 +13,7 @@ public class RClientCodegenTest {
         final RClientCodegen codegen = new RClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
