+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/rust/RustClientCodegenTest.java
@@ -13,7 +13,7 @@ public class RustClientCodegenTest {
         final RustClientCodegen codegen = new RustClientCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
