+++ modules/swagger-codegen/src/test/java/io/swagger/codegen/swift/SwiftCodegenTest.java
@@ -111,7 +111,7 @@ public class SwiftCodegenTest {
         final SwiftCodegen codegen = new SwiftCodegen();
         codegen.processOpts();
 
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), "true");
+        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
     }
 
     @Test
